+++ dimas/src/com/queryable.rs
@@ -133,7 +133,7 @@ where
 	///
 	/// # Panics
 	///
-	pub fn add(mut self) -> Result<()> {
+	pub fn add(self) -> Result<()> {
 		let collection = self.collection.clone();
 		let q = self.build()?;