+++ dimas/src/com/subscriber.rs
@@ -106,7 +106,7 @@ where
 	///
 	/// # Panics
 	///
-	pub fn add(mut self) -> Result<()> {
+	pub fn add(self) -> Result<()> {
 		let collection = self.collection.clone();
 		let s = self.build()?;