+++ dimas/src/lib.rs
@@ -1,7 +1,14 @@
 // Copyright © 2023 Stephan Kunz
 #![crate_type = "lib"]
 #![crate_name = "dimas"]
-#![warn(missing_docs)]
+//#![warn(missing_docs)]
+
+//! This crate is on [crates.io](https://crates.io/crates/dimas).
+//!
+//! `DiMAS` is tested on Windows (Version 10) and Linux (Ubuntu/Debian flavours) but should also run on `MacOS`.
+//!
+//! [DiMAS](https://github.com/dimas-fw/dimas/tree/main/dimas) follows the semantic versioning principle with the enhancement,
+//! that until version 1.0.0 each new version may include breaking changes, which will be noticed in the changelog.
 
 #[doc = include_str!("../README.md")]
 #[cfg(feature = "nightly")]
@@ -10,21 +17,23 @@ doc_comment::doctest!("../README.md");
 
 // region:    --- modules
 /// Primary module of `DiMAS` containing the Agent
-pub mod agent;
+mod agent;
 /// Communication
-pub(crate) mod com;
+mod com;
 /// Configuration
-pub mod config;
+mod config;
 /// Context
-pub mod context;
+mod context;
 /// Error handling
-pub mod error;
+mod error;
+/// Timer
+mod timer;
+// Helper functions
+//mod utils;
+
 /// Public interface of dimas.
 /// Typically it is sufficient to include the prelude with
 /// `use dimas::prelude::*;`
 pub mod prelude;
-/// Timer
-pub mod timer;
-// Helper functions
-//mod utils;
+
 // endregion: --- modules